lib: cbuf: Add flag to not use the event subsystem
CBUF is using the event framework to signal when the buffer has
available space to unblock waiting threads and internally the event
framework is using the thread_lock.

Unfortunately this is causing a deadlock when we want to access the CBUF
when the thread_lock is already taken by the kernel/scheduler, for
example during a context switch when we want to save the threads
information in the log (see KEVLOG_THREAD_SWITCH).

Since we are not going to use the blocking capabilities of CBUF, just
avoid using the event framework at all acting on a bool switch.

Signed-off-by: Carlo Caione <carlo.caione@nxp.com>

fixup! lib: cbuf: Add flag to not use the event subsystem

Signed-off-by: Xavier Roumegue <xavier.roumegue@nxp.com>
Signed-off-by: Antoine Bouyer <antoine.bouyer@nxp.com>