trace: Introduce tracelog subsystem
The tracelog subsystem is the framework in charge of saving the data
gathered by tracepoints in a circular buffer ready to be delivered to
the Linux side using ivshmem.

It is composed by a core and a set of modules (plugins) used to prepare
/ mangle the data to be saved in the logging buffer. Each module is
linked to a different type of data.

Currently we deal with three kind of data:
- Strings
- Generic binary data
- Data gathered by the LK debug system (KEVLOG_*)

Signed-off-by: Carlo Caione <carlo.caione@nxp.com>